chore: abstract readable stream transfer check
